requirements.txt
# Core Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3

# Database
SQLAlchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9  # Für PostgreSQL

# Authentication & Security
bcrypt==4.1.1
python-jose==3.3.0
passlib==1.7.4

# API & Serialization
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
flask-marshmallow==0.15.0

# AI Integration (wähle eine)
# openai==1.3.5
# anthropic==0.7.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
requests==2.31.0
python-dateutil==2.8.2

# Validation
email-validator==2.1.0
phonenumbers==8.13.26

# Logging & Monitoring
python-json-logger==2.0.7
sentry-sdk[flask]==1.38.0

# Task Queue (optional)
# celery==5.3.4
# redis==5.0.1

# File handling
python-magic==0.4.27

# Production Server
gunicorn==21.2.0
